An example unit test for the imageanalysis package.
I have created a simple unit test for the ImageController class which
tests a single path through the getImage method. I slightly restructured
ImageController to inject the ImageService class as this made testing
easier. This meant that ImageService need to be made stateless. I
haven't test run the program as I don't have the amazon and google keys.